integration test + extra getter to get data by level

+++ src/main/java/fact/it/game_service/controller/GameRestController.java
@@ -3,6 +3,7 @@ package fact.it.game_service.controller;
 import fact.it.game_service.model.Game;
 import fact.it.game_service.repository.GameRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.PostConstruct;
@@ -17,6 +18,7 @@ public class GameRestController {
 
     @PostConstruct
     public void fillDB(){
+
         if (gameRepository.count() == 0){
             gameRepository.save(new Game(1,"vraag 1", 1, 123.45, 567.89, "juist", "fout", "fout",4,4,"EXTRA_PotionVial"));
             gameRepository.save(new Game(2,"vraag 2", 1, 150, 50, "juist", "fout", "fout",5,0,"DEF_RoundShield"));
@@ -24,11 +26,11 @@ public class GameRestController {
             gameRepository.save(new Game(4,"vraag 4",2, 122.45, 566.89, "juist", "fout", "fout",0,2,"OFF_Pickaxe"));
         }
 
-        System.out.println(gameRepository.findGameByGameId(1).getQuestion());
+       System.out.println(gameRepository.findGameByGameId(1).getQuestion());
     }
 
     //get all questions
-    @GetMapping("/question")
+    @GetMapping("/questions")
     public List<Game> getAllQuestions(){
 
         return gameRepository.findAll();
@@ -39,7 +41,11 @@ public class GameRestController {
         Game q = gameRepository.findGameByGameId(id);
         return q;
     }
-
+    @GetMapping("/questionsbylevel/{level}")
+    public List<Game> getAllQuestionsByLevel(@PathVariable int level){
+        List<Game> q = gameRepository.findAllByLevelIs(level);
+        return q;
+    }
     //POST: question
     @PostMapping("/question")
     public Game createQuestion(@RequestBody Game question){
@@ -47,9 +53,9 @@ public class GameRestController {
     }
 
     //PUT:
-    @PutMapping("/question/{number}")
-    public Game updateQuestion(@RequestBody Game questionDetails, @PathVariable int number){
-        Game q = gameRepository.findGameByGameId(number);
+    @PutMapping("/question")
+    public Game updateQuestion(@RequestBody Game questionDetails){
+        Game q = gameRepository.findGameByGameId(questionDetails.getGameId());
         q.setQuestion(questionDetails.getQuestion());
         q.setLevel(questionDetails.getLevel());
         q.setX(questionDetails.getX());
@@ -66,9 +72,16 @@ public class GameRestController {
 
     //DELETE: question
     @DeleteMapping("/question/{number}")
-    public void deleteQuestion(@PathVariable int number){
+    public ResponseEntity deleteQuestion(@PathVariable int number){
         Game q = gameRepository.findGameByGameId(number);
-        gameRepository.delete(q);
+        if(q !=null) {
+            gameRepository.delete(q);
+            return ResponseEntity.ok().build();
+        }else{
+            return ResponseEntity.notFound().build();
+        }
+
+
     }
